+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.9.3] = 2026-xx-xx
+### Added
+- Cross sections can be linearized with `CrossSection.linearize`, and rotated and translated with
+  `CrossSection.rotate_translate`.
+- Combinations of cross section and camera config bounding box into wet and dry areas in the bounding box
+  can be found with a new function `CrossSection.get_bbox_dry_wet`
+### Changed
+### Deprecated
+### Removed
+### Fixed
+
+
 ## [0.9.2] = 2026-01-15
 ### Added
 ### Changed